Move the CDNPrioritizer to the fetchers code
