+++ src/Content_Steering.md
@@ -23,66 +23,66 @@ This separate file has its own syntax, semantic and refreshing logic.
 
 
 ```
-                                     ./src/parsers/SteeringManifest
-                            +----------------------------------+
-                            | Content Steering Manifest parser | Parse DCSM[1] into a
-                            +----------------------------------+ transport-agnostic steering
-                                    ^                            Manifest structure
-                                    |
-                                    | Uses when parsing
-                                    |
-                                    |
-                                    | ./src/transports
-                            +---------------------------+
-                            |        Transport          |
-                            |                           |
-                            | new functions:            |
-                            |   - loadSteeringManifest  | Construct DCSM[1]'s URL, performs
-                            |   - parseSteeringManifest | requests and parses it.
-                            +---------------------------+
-                                    ^
-                                    |
-                                    | Relies on
-                                    |
-                                    |
-   ./src/core/init                  | ./src/core/fetchers/steering_manifest
-   +---------+              +-------------------------+
-   |         | -----------> | SteeringManifestFetcher | Fetches and parses a Content Steering
-   |         |  Creates     +-------------------------+ Manifest in a transport-agnostic way
-   |         |                      ^                   + handle retries and error formatting
-   |         |                      |
-   |         |                      | Uses an instance of to load, parse and refresh the
-   |         |                      | Steering Manifest periodically according to its TTL[2]
-   |         |                      |
-   |         |                      |
-   |  Init   |                      | ./src/core/init/cdn_prioritizer.ts
-   |         |              +----------------+ Signals the priority between multiple
-   |         | -----------> | CdnPrioritizer | potential CDNs for each resource.
-   |         |  Creates     +----------------+ (This is done on demand, the `CdnPrioritizer`
-   |         |                     ^           knows of no resource in advance).
-   |         |                     |
-   |         |                     | Asks to sort a segment's available base urls by order of
-   |         |                     | priority (and to filter out those that should not be
-   |         |                     | used).
-   |         |                     | Also signals when it should prevent a base url from
-   |         |                     | being used temporarily (e.g. due to request issues).
-   |         |                     |
-   |         |                     |
-   |         |                     | ./src/core/fetchers/segment
-   |         |              +----------------+
-   |         | -----------> | SegmentFetcher | Fetches and parses a segment in a
-   +---------+  Creates     +----------------+ transport-agnostic way
-                                   ^           + handle retries and error formatting
-                                   |
-                                   | Ask to load segment(s)
-                                   |
-                                   | ./src/core/stream/representation
-                            +----------------+
-                            | Representation | Logic behind finding the right segment to
-                            |    Stream      | load, loading it and pushing it to the buffer.
-                            +----------------+ One RepresentationStream is created per
-                                               actively-loaded Period and one per
-                                               actively-loaded buffer type.
+               /parsers/SteeringManifest
+      +----------------------------------+
+      | Content Steering Manifest parser | Parse DCSM[1] into a
+      +----------------------------------+ transport-agnostic steering
+              ^                            Manifest structure
+              |
+              | Uses when parsing
+              |
+              |
+              | /transports
+      +---------------------------+
+      |        Transport          |
+      |                           |
+      | new functions:            |
+      |   - loadSteeringManifest  | Construct DCSM[1]'s URL, performs
+      |   - parseSteeringManifest | requests and parses it.
+      +---------------------------+
+              ^
+              |
+              | Relies on
+              |
+              |
+              | /core/fetchers/steering_manifest
+      +-------------------------+
+      | SteeringManifestFetcher | Fetches and parses a Content Steering
+      +-------------------------+ Manifest in a transport-agnostic way
+              ^                   + handle retries and error formatting
+              |
+              | Uses an instance of to load, parse and refresh the
+              | Steering Manifest periodically according to its TTL[2]
+              |
+              |
+              | /core/fetchers/cdn_prioritizer.ts
+      +----------------+ Signals the priority between multiple
+      | CdnPrioritizer | potential CDNs for each resource.
+      +----------------+ (This is done on demand, the `CdnPrioritizer`
+             ^           knows of no resource in advance).
+             |
+             | Asks to sort a segment's available base urls by order of
+             | priority (and to filter out those that should not be
+             | used).
+             | Also signals when it should prevent a base url from
+             | being used temporarily (e.g. due to request issues).
+             |
+             |
+             | /core/fetchers/segment
+      +----------------+
+      | SegmentFetcher | Fetches and parses a segment in a
+      +----------------+ transport-agnostic way
+             ^           + handle retries and error formatting
+             |
+             | Ask to load segment(s)
+             |
+             | /core/stream/representation
+      +----------------+
+      | Representation | Logic behind finding the right segment to
+      |    Stream      | load, loading it and pushing it to the buffer.
+      +----------------+ One RepresentationStream is created per
+                         actively-loaded Period and one per
+                         actively-loaded buffer type.
 
 
 [1] DCSM: DASH Content Steering Manifest
