+++ src/core/fetchers/segment/segment_fetcher_creator.ts
@@ -15,12 +15,15 @@
  */
 
 import config from "../../../config";
+import Manifest from "../../../manifest";
 import {
   ISegmentPipeline,
   ITransportPipelines,
 } from "../../../transports";
-import CdnPrioritizer from "../../init/cdn_prioritizer";
+import { CancellationSignal } from "../../../utils/task_canceller";
 import { IBufferType } from "../../segment_buffers";
+import CdnPrioritizer from "../cdn_prioritizer";
+import SteeringManifestFetcher from "../steering_manifest";
 import applyPrioritizerToSegmentFetcher, {
   IPrioritizedSegmentFetcher,
 } from "./prioritized_segment_fetcher";
@@ -89,9 +92,19 @@ export default class SegmentFetcherCreator {
    */
   constructor(
     transport : ITransportPipelines,
-    cdnPrioritizer : CdnPrioritizer,
-    options : ISegmentFetcherCreatorBackoffOptions
+    manifest : Manifest,
+    options : ISegmentFetcherCreatorBackoffOptions,
+    cancelSignal : CancellationSignal
   ) {
+    const steeringManifestFetcher = transport.steeringManifest === null ?
+      null :
+      new SteeringManifestFetcher(transport.steeringManifest,
+                                  { maxRetryOffline: undefined,
+                                    maxRetryRegular: undefined });
+    const cdnPrioritizer = new CdnPrioritizer(manifest,
+                                              steeringManifestFetcher,
+                                              cancelSignal);
+
     const { MIN_CANCELABLE_PRIORITY,
             MAX_HIGH_PRIORITY_LEVEL } = config.getCurrent();
     this._transport = transport;
